# perf: Make `VISITOR_KEYS` etc. faster to access