--- a/packages/babel-types/package.json
+++ b/packages/babel-types/package.json
@@ -25,8 +25,7 @@
   },
   "dependencies": {
     "@babel/helper-string-parser": "workspace:^",
-    "@babel/helper-validator-identifier": "workspace:^",
-    "to-fast-properties": "condition:BABEL_8_BREAKING ? ^3.0.0 : ^2.0.0"
+    "@babel/helper-validator-identifier": "workspace:^"
   },
   "devDependencies": {
     "@babel/generator": "workspace:^",