--- a/packages/babel-types/src/definitions/index.ts
+++ b/packages/babel-types/src/definitions/index.ts
@@ -1,4 +1,3 @@
-import toFastProperties from "to-fast-properties";
 import "./core.ts";
 import "./flow.ts";
 import "./jsx.ts";
@@ -29,15 +28,16 @@ import { DEPRECATED_ALIASES } from "./deprecated-aliases.ts";
 });
 
 // We do this here, because at this point the visitor keys should be ready and setup
-toFastProperties(VISITOR_KEYS);
-toFastProperties(ALIAS_KEYS);
-toFastProperties(FLIPPED_ALIAS_KEYS);
-toFastProperties(NODE_FIELDS);
-toFastProperties(BUILDER_KEYS);
-toFastProperties(DEPRECATED_KEYS);
+// `%ToFastProperties` no longer improves performance, so we use another approach.
+const _VISITOR_KEYS = { ...VISITOR_KEYS };
+const _ALIAS_KEYS = { ...ALIAS_KEYS };
+const _FLIPPED_ALIAS_KEYS = { ...FLIPPED_ALIAS_KEYS };
+const _NODE_FIELDS = { ...NODE_FIELDS };
+const _BUILDER_KEYS = { ...BUILDER_KEYS };
+const _DEPRECATED_KEYS = { ...DEPRECATED_KEYS };
 
-toFastProperties(PLACEHOLDERS_ALIAS);
-toFastProperties(PLACEHOLDERS_FLIPPED_ALIAS);
+const _PLACEHOLDERS_ALIAS = { ...PLACEHOLDERS_ALIAS };
+const _PLACEHOLDERS_FLIPPED_ALIAS = { ...PLACEHOLDERS_FLIPPED_ALIAS };
 
 const TYPES: Array<string> = [].concat(
   Object.keys(VISITOR_KEYS),
@@ -46,17 +46,17 @@ const TYPES: Array<string> = [].concat(
 );
 
 export {
-  VISITOR_KEYS,
-  ALIAS_KEYS,
-  FLIPPED_ALIAS_KEYS,
-  NODE_FIELDS,
-  BUILDER_KEYS,
+  _VISITOR_KEYS as VISITOR_KEYS,
+  _ALIAS_KEYS as ALIAS_KEYS,
+  _FLIPPED_ALIAS_KEYS as FLIPPED_ALIAS_KEYS,
+  _NODE_FIELDS as NODE_FIELDS,
+  _BUILDER_KEYS as BUILDER_KEYS,
   DEPRECATED_ALIASES,
-  DEPRECATED_KEYS,
+  _DEPRECATED_KEYS as DEPRECATED_KEYS,
   NODE_PARENT_VALIDATIONS,
   PLACEHOLDERS,
-  PLACEHOLDERS_ALIAS,
-  PLACEHOLDERS_FLIPPED_ALIAS,
+  _PLACEHOLDERS_ALIAS as PLACEHOLDERS_ALIAS,
+  _PLACEHOLDERS_FLIPPED_ALIAS as PLACEHOLDERS_FLIPPED_ALIAS,
   TYPES,
 };