<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | `Fixes #1, Fixes #2` <!-- remove the (`) quotes and write "Fixes" before the number to link the issues -->
| Patch: Bug Fix?          |
| Major: Breaking Change?  |
| Minor: New Feature?      |
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
This fixes potential https://github.com/babel/babel/pull/16873#issuecomment-2422689587

Also improved access speed for some properties.

before
```
PS F:\babel\benchmark\babel-traverse> node .\real-case.mjs --only-current
babel-traverse/real-case.mjs babel-parser-express.ts @ current: 299 ops/sec ±1.82% 150 runs (3.348ms)
babel-traverse/real-case.mjs jquery-3.6.js @ current: 87.09 ops/sec ±2.4% 44 runs (11ms)
babel-traverse/real-case.mjs ts-checker.mjs @ current: 10.94 ops/sec ±2.32% 30 runs (91ms)
babel-traverse/real-case.mjs ts-parser.mjs @ current: 81.14 ops/sec ±2.77% 41 runs (12ms)
babel-traverse/real-case.mjs ts-parser.ts @ current: 54 ops/sec ±2.2% 30 runs (19ms)
babel-traverse/real-case.mjs typescript-5.6.2.js @ current: 2.73 ops/sec ±1.61% 30 runs (366ms)
```
after
```
PS F:\babel\benchmark\babel-traverse> node .\real-case.mjs --only-current
babel-traverse/real-case.mjs babel-parser-express.ts @ current: 313 ops/sec ±2.07% 157 runs (3.197ms)
babel-traverse/real-case.mjs jquery-3.6.js @ current: 102 ops/sec ±2.52% 52 runs (9.759ms)
babel-traverse/real-case.mjs ts-checker.mjs @ current: 11.68 ops/sec ±1.7% 30 runs (86ms)
babel-traverse/real-case.mjs ts-parser.mjs @ current: 85.35 ops/sec ±3.97% 43 runs (12ms)
babel-traverse/real-case.mjs ts-parser.ts @ current: 58.79 ops/sec ±2.33% 30 runs (17ms)
babel-traverse/real-case.mjs typescript-5.6.2.js @ current: 2.85 ops/sec ±2.43% 30 runs (351ms)
```


This is not due to the `to-fast-properties` package, as `%HasFastProperties` shows that it correctly makes the object fast.
I suspect the original optimization here is being lost due to some changes in v8, even if I use the `%ToFastProperties` provided by v8, it still performs the same as without special optimizations.

Take the parser as an example
When we go from 19 to 20 options, the return value of `%HasFastProperties` changes from true to false, and the overall performance drops by 10%.
After we call `%ToFastProperties`, the return value of `%HasFastProperties` changes from false to true, but the performance does not return to the previous state.
Until we create a new fast object using `{...obj}`, performance will return to previous state.