@liuxingbaoyu 2024-10-19
I made a minimal reproduction. We can find that `%ToFastProperties(obj)` and `{...obj}` are both fast objects, but the performance is different.

```
PS F:\v8-fast-object> node --allow-natives-syntax main.mjs fast     
fast: 5.823s
PS F:\v8-fast-object> node --allow-natives-syntax main.mjs toFast
toFast: 6.719s
PS F:\v8-fast-object> node --allow-natives-syntax main.mjs cloneFast
cloneFast: 5.777s
```
```js
const object1 = {
  a1: null,
  a2: [],
  a3: 0,
  a4: "script",
  a5: undefined,
  a6: true,
  a7: true,
  a8: true,
  a9: true,
  a10: true,
  a11: true,
  a12: true,
  a13: true,
  a14: true,
  a15: true,
  a16: true,
  a17: true,
  a18: true,
  a19: true,
  a20: true,
};

let x;

function test(opts, name) {
  console.time(name);

  for (let i = 0; i < 1e10; i++) {
    x = opts.a1;
  }
  console.timeEnd(name);
}

const name = process.argv[2];
switch (name) {
  case "fast":
    test(object1, name);
    break;
  case "slow":
    const object2 = {};
    for (const key of Object.keys(object1)) {
      object2[key] = object1[key];
    }

    test(object2, name);
    break;
  case "toFast":
    const object3 = {};
    for (const key of Object.keys(object1)) {
      object3[key] = object1[key];
    }
    %ToFastProperties(object3);

    test(object3, name);
    break;
  case "cloneFast":
    const object4 = { ...object1 };

    test(object4, name);
    break;
}
```