@liuxingbaoyu 2024-10-20
Oops, my description in the PR was wrong.
Because we are compiling to node 6, `{...obj}` will become `Object.assign({}, obj)`.
This will make objects slow. So `@babel/parser` theory is correct, but it doesn't actually work in Babel 7. And `@babel/types` theory is the opposite, and it actually works. 😅